Add HA tests for s390x on zKVM without support server
This moves several functions performed in the support server
on HA tests such as the mutex and barrier creation, LUN list
generation and name resolution from the support server to
node 1 of the cluster setup when not using the support server.

As a result, HA test schedule is modified on such scenario.

Also included is the definition of new default timeout values
on lib/hacluster used to scale the timeout of several method,
and the creation of new variable NFS_SUPPORT_DIR and
NFS_SUPPORT_SHARE used as a shared communication method
between nodes to share the LUN list file and the /etc/hosts.