New test: installation on remote multipathed disk
During initial installation the SUT imports a multipathed iSCSI target disk
from a parallel supportserver job and proceeds to install onto it.

Once the initial installation is complete and the client is ready to reboot
the supportserver adds the new kernel and initrd to its PXE server
configuration (custom_pxeboot.pm).  After this the client is supposed
to boot this kernel and initrd via PXE.  The initrd would then re-import
the remote multipathed system disk and proceed from there.

__FIXME__
   As of now this works only on SLES-15 SP1.  Only there the partitioner
   prefers (for some unknown reason) the imported disk over the local one
   for the system disk role.

   For other products either the partitioner would have to be modified
   accordingly or the default local disk of the SUT would have to be taken
   away in the first place (NUMDISKS set to 0. This is currently NOT
   supported, though).