Allow building custom kernels for testing
Some scenarios require us to be able to not only test downstream
kernels, but also upstream development trees with our QA hardware.
In this case we need to download, configure, build, install and boot
custom kernels from arbitrary GIT trees defined in the test settings.

The URL to the kernel tree is specified in KERNEL_GIT_TREE, the branch
to checkout in KERNEL_GIT_BRANCH. If KERNEL_GIT_TREE is set, this is
module is loaded and scheduled after installation.

In case kernel building fails, the complete output will be uploaded to
openQA.

Signed-off-by: Michael Moese <mmoese@suse.de>