Use regexp against STDOUT instead of screen check
STDOUT is just text and a screen check is not as reliable because a font
change, a resolution change would need a new needle.
Furthermore, different needles have to be maintained for each product.
script_output and validate_script_output should be used instead of
check_screen and assert_screen if possible.