Add support for graphical installer when using ipxe_install
Sometimes, for a machine there is no autoyast profile available yet,
autoyast is broken or we just want to make sure the installer is working
on bare metal.
In this case, we should start the installer as we would for the
traditional installations over IPMI so we can use the graphical
installer with all that comes with it.

This change will allow to use ipxe installations without autoyast by
unsetting the AUTOYAST and AUTOYAST_PREPARE_PROFILE variables.

Signed-off-by: Michael Moese <mmoese@suse.de>