Wicked: Remove tracepath from basic suite t08
This tracepath makes sense when we have a setup similar
to the legacy test suite:
https://github.com/openSUSE/wicked-testsuite/blob/master/README.setup#L6

We have a different setup in OpenQA and we would need to force REF
as a router to forward packets to the "outside" world. After talking
to Wicked team, the aim of this test is just to verify that Wicked
can configure 2 interfaces, so there is no need to test routing here.