Replace old IP addresses in HA configuration on migrations
When testing SLES+HA migrations without support server and with
a cluster configured in unicast, IP addresses of the original
cluster could have changed and cluster will fail to start in
migrated nodes. This commit determines both the old and new IP
address, and replaces the old ones for the new ones in the HA
configuration files, before restarting the cluster.

Workaround is applied only in node1 and then files are synced
with csync, so clusters with more than 2 nodes are supported.