supportserver PXE: add custom boot menu entry
Suppose one of the supportserver's disks constitutes the system disk of
a client (e.g. via iSCSI). This extension to the PXE service enables
the supportserver to retrieve the kernel/initrd pair from the
installation on this disk and create an additional PXE boot entry
from it. The client may then use this for booting.