New SLES-15 SP1 supportserver
The currently available supportserver is SLES-12 SP3. This is an old
product. E.g. it was found to be unable to properly mount/umount SLE-15
btrfs: this ability is needed for the upcoming installation test of
a SLES-15 client on a (remote) supportserver disk.

The available documentation/config on supportserver generation does not
really provide much assistance on how to proceed. Therefore this commit
provides a generator script and a tentative autoyast file from which a
SLES-15 SP1 supportserver disk image can be generated in a one-time job run.

FIXME: the autoyast file specifies internal SUSE repositories for
simplicity. In order to utilize the official repositories sections
for the needed registrations would have to be included.