New test: installation on remote multipathed disk
During initial installation the SUT imports a multipathed iSCSI target disk
from a parallel supportserver job and proceeds to install onto it (the
testsuite must set SELECT_FIRST_DISK for this).

Once the initial installation is complete and the client is ready to reboot
the supportserver adds the new kernel and initrd to its PXE server
configuration (custom_pxeboot.pm).  After this the client is supposed
to boot this kernel and initrd via PXE.  The initrd would then re-import
the remote multipathed system disk and proceed from there.

_Note:_  For client installations >= SLE-15 (or equivalent) the supportserver
also needs to be SLE-15 or newer, since older supportservers are unable to
properly access the client's new (btrfs) rootfs and kernel/initrd.