Delete unused code pre_bootmenu_setup
I checked tests on both o3 and osd and could not find the combination of
IPXE with bootloader or mediacheck. IPXE is used in special, different
test setups which does not rely on pre_bootmenu_setup at all. Only the
combination of the bootloader module together with BOOT_HDD_IMAGE might
be used so this commit moves that check directly into the bootloader
code. Obviously mediacheck is also not used with BOOT_HDD_IMAGE and
!UEFI and I would doubt this could even work.