Add workaround to select first item in module list
The workaround is added because description for "Basesystem Module"
is not appeared when the module is selected for the first time.

On s390x the only way to ensure the module is selected is to use
description for needle match, because item highlighting is not
recognized by openQA. The same needle matches in both cases - when
item is highlighted and not.

Related ticket: [poo#59939](https://progress.opensuse.org/issues/59939)