Add locking support for baremetal machines
Sometimes, we really need to prevent openQA from using the baremetal
test hardware. This is especially the case, when we need the specialized
hardware in those machines for test development or analyzing failures
etc. Without the possibility to lock the machines from openQA usage, a
new build would kick the user out of the machines and re-deploy them.

In order to avoid this, there is a support service
(https://github.com/frankenmichl/baremetal_support) which allows us to
do this locking.
In case a machine is locked, tests may fail. If noone interacts with the
machines, the locking will always succeed. The service does not store
data persistently, so in case of a service crash the machines would
default to an unlocked state.

Signed-off-by: Michael Moese <mmoese@suse.de>