Fix different handling for gdm password depending on product version
This should fix the password handling in the gdm login prompt depending on
product versions. In general "is_tumbleweed" should be avoided to handle new
features which are expected to come to SLE and Leap eventually anyway.
This problem was introduced in e288e39c.

This now handles older products as the exception as it should be, not the
other way around.

Related progress issue: https://progress.opensuse.org/issues/60332