Wrapper for using runltp-ng on baremetal
Minimal OpenQA test module for running the runltp-support scripts which
in-turn wrap runltp-ng and expect. To begin with this just targets the
syzkaller reproducers as most LTP tests run fine in OpenQA, however the
reproducers can crash the host (worker) in addition to the VM being tested.

Also there are 1000's of them and OpenQA struggles to handle logging this many
tests. For the same reason the OpenPOSIX tests will benefit from being run
like this, but they are not supported yet.