Add test suit for GraphicsMagick
GraphicsMagick is a package to manipulate graphics in command line
and in a graphic view similar to ImageMagick.

The official webpage is: www.graphicsmagick.org/

This test has two parts. First a console script with 106 unitary
tests  and secondly 4 needle tests. The script uses the graphicsmagick
comparison tool to check the success of the tests. This way increase
the speed vs using needles. The needles are used only
on tests that are visual because the tool opens windows (gm display)

Step 1: executes an script with these sections:

- Test 1. Comparison tool
- Test 2. Conversions
- Test 3. Resize, rotate, crop, flip
- Test 4. Special effects and modifications
- Test 5. Transparencies
- Test 6. Montage of image thumbnails
- Test 7. GIF animations
- Test 8. Composite images
- Test 9. Draw shapes and text on a image
- Test 10. Big images
- Test 11. Batch operations
- Test 12. Mogrify
- Test 13. Scripting

Note: All image results are stored with the number of the number of the
test to help debugging

Step 2: executes some visual tests checked by needles:

- Test 1. Open an image
- Test 2. Open an image specifying the geometry of the window and position
- Test 3. Open a list of images
- Test 4. View the histogram of an image