Fix 'linux_boot_entry' for SLE15+
By default in 'handle_grub' function, the '$linux_boot_entry' variable is
set to 14 lines, but this doesn't work in sle15+, as 'load_video' Grub
function is called in 15+ OSes.
So '$linux_boot_entry' should be set to 15 in that case.

This commit also add a space before typing '$boot_params', as a space is
not always set at the end of the line and having 2 spaces doesn't hurt.