Enable remote installation over VNC for openSUSE
yaml schedulers for the controller and the target for opensuse.
i have already created the test suites with the corresponding yaml_schedule.
The NETBOOT and INSTALL_SOURCE are ommitted from the configuration
as the DVD flavor can be used.
I tried to take away the HDD_1 from the test_suite and move it in the
yaml but the job fails reporting that the image cant found( searching in
/var/lib/openqa/pool/3/raid/ somehow)
I had to terminate the yast2 manually after the mutex.