HPC: Package versions and integration tests
This patch is introducing very basic sanity checks for white-listed
packages. The justification is that the HPC project is often just
simply bumping rpms from version x to version y. It makes sense to
introduce some sanity, quick checks on those versions and see if they
are as expected in the end product. On top of that the patch is taking
care of some very simple, single machine checks for the listed rpms,
so that it can be ensure that very basic function of the given rpm
is working