Set DESKTOP=gnome explicitly for the btrfs and ext4 tests
During the review we have found failing tests on zVM because early
return somehow didn't work for the setup and tests started to fail.
However, there was a second issue we missed as DESKTOP was changed to
textmode for many of the scenarios we had.
I was not able to figure out why `SCC_ADDONS` setting got changed
between last good and failing ones, but as a result we started selecting
different modules and end up with text mode.
With this change, we set DM explicitly to have same behavior as in build
89.1.

See [poo#61964](https://progress.opensuse.org/issues/61964).