HPC: ensure correct hostname
Current multimachine/HPC tests rely on using support server, so that
for instance DHCP is used for assigning hostnames etc. This overall
might be problematic for some specific HPC tests, where correct set-up
of a private network is a requirement and not the goal of the tests.
For this reason it make sense to gradually decouple cluster provisioning
form the test execution.

This patch is an introduction to a new hpc specific lib which could
eventually ensure specific configuration for creating private network
and thus provision HPC cluster are done as expected

Initially the plan was to completly remove support config from cluster
provisioning however as of now slurm updates are getting no acceptance
so this intermediate commit is merged