Partial revert of "wicked: enable cleanup_before_shutdown"
With PR #9142 "wicked: enable cleanup_before_shutdown" some
cleanup logic was introduced. It appears that this cleanup
logic is breaking too many things and partially no-op :

1. Delete of '/etc/hostname' is breaking assumption of
many tests about hostname 'susetest'.
2. Cleanup in /var/run folder does not make any sense
because /var/run is symlink to /run which is tmpfs

Based on this [1] is limited to scenario which preparing
image for wicked tests where absent of this file does not
cause any troubles. And [2]  is fully reverted because it
is pure waste of time