Set SYSTEM_ROLE to textmode for SLES4SAP 15+ only on qemu
In SLES4SAP 15+ tests, SYSTEM_ROLE is set to textmode when
DESKTOP=textmode. This is useful to test the system in textmode, but
does not allow the tests of SLES4SAP in textmode with the default SLES
for SAP Applications system role. Tests in textmode with the default
SLES for SAP Applications system role are necessary when testing SLES4SAP
over the spvm or pvm_hmc back ends. This commit limits the change of the
SYSTEM_ROLE variable to textmode when DESKTOP=textmode only for SLES4SAP
tests over qemu where a graphical environment is available.