sudo: Stop abusing /proc/sys/vm/drop_caches
In newer kernels (>= 5.4), the drop_caches proc file has been turned into a
write-only file. As such, reading from it is no longer permitted.

https://github.com/torvalds/linux/commit/204cb79ad42f015312a5bbd7012d09c93d9b46fb

In fact, we don't need drop_caches, as we simply abused this file before. Let's simple
write to /run/openqa_sudo_test, which is also only writable by root. As a bonus,
we always expect to read from the file what we just wrote to it, making it a bit more
obvious what to expect.

https://progress.opensuse.org/issues/62168