Skip unnecessary serial getty disable on more recent products
We are currently applying some workarounds to make access to the serial port
from tests more reliable. According to
https://bugzilla.suse.com/show_bug.cgi?id=1112108#c3 this should not be
necessary anymore with Linux => 4.20.4 but we keep the workaround for older
products. This is done in a conservative way as older products might have
received the same bugfix by backport but we still apply the workaround to be on
the safe side.

Less workarounds make the tests cleaner and faster.

Related progress issue: https://progress.opensuse.org/issues/54293