await_install: Prevent endless loop on false-positive needle matches
As happened in
https://openqa.opensuse.org/tests/latest?test=kde_live_upgrade_leap_42.3&version=Tumbleweed&arch=x86_64&machine=64bit-2G&distri=opensuse&flavor=KDE-Live
a black screen was incorrectly matched as "screenlock" causing the
internal timeout counter to never decrease until the test job is aborted
as incomplete without any logs by openQA. This commit changes behaviour
to always decrease the timeout regardless of the match result. This
simplifies the logic and prevents an endless loop. The only other
functional impact is that the overall timeout time is reduced by a
negligible 30s for each correct match.

Related progress issue: https://progress.opensuse.org/issues/48899