HPC: Deal with Service Pack 0 while migrating HPC
While checking for correct number of migration targets, while initial
platfrom to migrate from is the service pack 0 of given major code
stream version, the returned $current_version value is that of a major
code stream version. So for SLE15 SP0 the value is 15. That makes the
check of correct number of possible migration targets incorrect