publiccloud: terraform_destroy() - avoid on upload
There is a special handling for publiccloud_upload_img test. But the
check fails, if a such a test was triggered with openqa-clone-job*
This change check if the TEST variable begins with
`publiccloud_upload_img`, which should do it for now.

In general, I would prefer a solution where we do not rely on a test name.
But I don't see the need to invest more for now.