Add soft-failure for the remote vnc installation on openSUSE
We have a kernel bug in TW which is visible in openQA, but we were not
able to catch it manually. System is able to boot normally, if we force
reboot by pressing ctr-alt-delete, so that will be used as a workaround
for now.

See [poo#63247](https://progress.opensuse.org/issues/63247).