Enhance code checking rules
we have a tool, test_deleted_renamed_referenced_modules.
It will check a module be rename/deleted if still is in used.

But there is a bug when rename a module. the old name is a substring of
new name, it will report check failed.

For example, the old module is 'cpu_bugs/xen_pv.pm', new module is
'cpu_bugs/xen_pv_hvm.pm', the tool will check 'cpu_bugs/xen_pv' string
to determine the old one if still is in used.
and then it will match the new one, because it is substring of the new
one. the tool will judge the old one still is in used. break code
checking.