Reuse existing root-ssh for root-console
Fix poo#59843: If is `select_console 'root-console'` called several
times on remote backend, second use will fail. It works properly when
it runs only once. When is `root-console` selected for the first time,
`root-ssh` is used during console initiation. When is `root-console`
selected after, it will fail, the switch to `root-ssh` will not happen.