Enable IUCV console via linuxrc parameter
As suggested by Berthold in
https://bugzilla.suse.com/show_bug.cgi?id=1156053#c32

With IUCV on current kernels, there are two issues. First ist the one
with HARDENED_USERCOPY, the other one is, that current kernels do not
provide a hvc iucv console by default anymore. Please try to add
hvc_iucv=8 to the kernel command line to make sure that this is not the
reason for the failure.