publiccloud: Improve timeout handling for azure login
Azure login sometimes takes more time. With `script_run()`, we faced the
problem that the previous command timed out, but we didn't checked for
valid shell prompt. So we messed up the shell parsing and sometimes used
the result of previous command.
With this mess shell handling, we called `az login` sometimes twice,
even if we already got valid login and it appears that the second call
failed :/.