Set initiator to automatic startup after configuration
in the service widget it was selected to "restart" the service with the "on_boot" after reboot option.
But then the test checked that the iscsid.socket was active.
In that case "on_demand" after reboot should be selected.

Then, the test checked active sessions with iscsiadm, but, the targets
were configured to be connected manually, so after restarting the services there was no sessions.
Setting this to autoatic fixes the issue with the test

Testing with yast2-iscsi-client-4.2.5-1, just wait for the fix to reach latest build.