publiccloud: Define img_proof test list it the module
Till now, we defined the variable PUBLIC_CLOUD_IMG_PROOF_TESTS in each
testsuite, which needed to defined 6 suites.

With this change we make the decission which tests are running for which
FLAVOR inside the test-module.

Also fixing the problem, that we run ec2_byos tests on EC2-CHOST-BYOS
image, but this image doesn't have ec2metadata installed bsc#1166890.