Use $include instead of !include as is invalid yaml
We have used '!include' tag for referencing another yaml file in the
test_data section of the schedule. If not quoted this is invalid yaml
format and worked due to a bug in YAML::Tiny.

Using `$include` doesn't introduce any of these issues.

See [poo#64755](https://progress.opensuse.org/issues/64755).