LTP: Add dependencies for net.tcp_cmds
telnet-server and xinetd are "maybe_deps" because they are available
only Tumbleweed (not on SLES).

Also merge 3 arrays used for zypper_call() which speedup executing a bit
(calling zypper_call() only once instead of 3 times).

This code is not run as the default, so it's just mirroring changes in
package.