LTP: Move kernel baremetal setup to main_common.pm
load_baremetal_tests() call for LTP_BAREMETAL && INSTALL_LTP
was moved to main_common.pm which enables this fix also for openSUSE.

This required to move also load_baremetal_tests(), which in turn causes
main_ltp.pm to load some modules from main_common.pm and
products/sle/main.pm to load load_baremetal_tests().

Although IPMI is not yet used in o3, be prepared for future.
But this code is reused for s390x installation on o3, which also needs
to use installer (all tests needs to be chained to install_ltp with
START_DIRECTLY_AFTER_TEST, because on o3 there is only z/VM which does
not handle qcow).

Fixes: d83486058 ("LTP: Introduce bare metal tests")