Implement wpa_supplicant test
Adds a test for wpa_supplicant. The wpa_supplicant test is a
self-sufficient zip archive that gets extracted and executed.

The wpa_supplicant test creates two virtual wlan interfaces and tests
the wlan functionality by setting up a hotspot via hostapd.
wpa_supplicant is then tested, if able to obtain a network list and
able to connect to an open and encrypted network. Lastly we test if
wicked is able to get a ip address via dhcp on the wlan interface.