Automate Salt Formulas with Forms support in AutoYaST
The new test suite will start an AY installation and after the main
installation, it passes the control to the salt. This happens having the
configuration_management section defined in the AY. To exercise this we
use masterless configuration, to avoid the use of a helper server acting as
the salt server which serves the requiring files.

In masterless mode the configuration_management needs to point to a tar file
which containing what you would put under /srv/salt in the server/client mode
(that's it, a 'salt' and a 'pillar' directories)

The tar used, is taken by the example provided by igonzalezsosa
with a small modification.

I use SALT_FORMULAS_PATH to point to this tar, which is stored in data and i used regular
approach we use to retrieve add access files under /data.
New module is created to verify the impact of the motd configuration.