Fixes yast2_lan_restart restart checks
Make the yast2_lan_restart compliant with the recent change which the reload
is the expected behaviour instead of a restart.

Minor changes and refactoring replacing certain statements with existing
functions and error output of the assertions.

Now instead of checking if the journal.log has found any of the filters,
we try to match status against the journal.log context based of the
expectations.

Apply suggestions from code review

Accept comment change

Co-Authored-By: Anton Smorodskyi <23258171+asmorodskyi@users.noreply.github.com>