Fix autoyast profile handling when used as an asset
Due to changes in openQA we now need to upload assets with unique name
of the files as openQA doesn't handle this anymore. Also, ASSET_*
variables are now expanded to the full path of the asset and do not
require any further guessing of the path.

See [poo#65229](https://progress.opensuse.org/issues/65229).