Add IOMMU commandline options to autoyast profile
In the future, we plan for virtualized testing with our mlx_con5
machines. As one preparing step, we need to enable a few command line
options to enable the IOMMU.
The option pci=realloc is usually only needed when we have a buggy BIOS.
Until it is possible to update the firmware on sonic and tails we should
just go with this option.

Signed-off-by: Michael Moese <frankenmichl@posteo.de>