Wicked: Add IPv6 library and example of usage with t01
IPv6 suite has a lot of should/should_not actions depending
on what is the expected result. Having only a function
to check all the cases is too messy. This way, we can have
control of all the steps and add functionality to a specific
step without adding too much code.