HPC: Rasdaemon test extensions
Over the lifecycle of this sanity tests the content was changed where
some test ideas were removed - namely error injection. This patch is
restoring that original error injection, for SLE15-SP2, while it
preserves all added basic sanity checks. Moreover the test description
is changed to clearly indicate what are the limitations for the test,
why it does exists as the openQA test module and what it does.
As such the test for rasdaemon in virtualized env. has very limitted
purpose and is meant only as a sanity check for the isolated package to
see if package is installable and if the daemon itself can run. As the
daemon is run with the -r option it writes data to the databese, so the
test is checking if basic information can be retrived from the db, also
if the injected error can be retrived