Expose testsuite options to openQA
When running the InfiniBand testsuite, it may be required to skip
certain tests or modify the configuration option. This is now possible
by adding variables to the master testsuite. The variables are
documented inside the source code, and the entire documentation was
updated to be fully up-to-date.
With these changes, all command line options offered by the testsuite is
now available to be modified on openQA testsuite level.

Signed-off-by: Michael Moese <frankenmichl@posteo.de>