Add a network bridge to default installation
In preparation for deploying virtual machines from test code on the
baremetal machines, we should deploy new installations with a bridged
network by default. We won't have any negative impact, but deploying
virtual machines is easier, if we don't need to reconfigure the
interface we use for connecting to the machine.

Signed-off-by: Michael Moese <frankenmichl@posteo.de>