Enable gnome installation with /home partition encrypted
We perform an AutoYast installation creating a separate partition
for home and we encrypt it. In additional we select Workstation Extension
addon and we install gnome-x11 pattern.
As occurs with encrypted boot partition the installation module stops
after the grub and we handle the decryption in the next module.
We verify the scenario at the end reusing the verify_separate_home
and testing the clone function.