Validate YAML schedules against JSON Schema
Issue: https://progress.opensuse.org/issues/66230

This makes sure that the schedule files contain the correct structure.

Otherwise it can happen that it contains things like:

    schedule:
    - {{foo}}

Which is valid YAML, but it's actually a YAML mapping, short for:

    schedule:
    - { { foo: null }: null }

and it must be quoted.
The JSON Schema validation will make sure it is always a string.