nautilus-open_ftp: behave more like a user would
Most users won't click an item in the location bar and then chose to perform
the action (in this case unmount) using a keyboard shortcut.
It's much more realistic for the user to click 'unmount' in this case.

This fixes the test insofar as Nautilus 3.36.2 fixed an earlier broken behavior:
  When a user selected Location 1, scrolled 60% down, then selected Location 2,
  it could happen that the newly loaded Location 2 showed at 60% scrolled down
  Newly, nautilus scrolls to the top, which in turn has the side effect that
  the focus is back in the files area, so alt-f10 would not trigger the menu
  on the item this test expects it to happen on

https://progress.opensuse.org/issues/66331