networkd: Specify grep as a required package to be installed
The test code explicitly calls grep and it is thus our responsibility
to ensure the tool to be present.

In the past, grep used to be pulled in by every package that had %fillup_prereq
defined in the preamble (in the specific case of the networkd container, this
used to the case for issue-generator). With the fillup_prereq macro being adjusted
to the current situation, diffutils, and as a result grep (required by diffutils)
are no longer present in the dep chain

https://progress.opensuse.org/issues/66373