Verify initrd contents after kernel livepatch package installation
A kernel livepatch package's %post is supposed to update the target
kernel's corresponding initrd, so that the livepatch can be loaded during
boot. This had failed at times in the past, so make update_kernel verify
that the initrd has been updated after the update to the kernel livepatch
package from $INCIDENT_REPO. To this end, check that all kernel modules
provided from the kernel livepatch package have been made available in the
initrd and that the livepatching dracut module has been enabled there.

As this functionality will also be needed for future tests, wrap it in a
function verify_klp_pkg_installation() exported from the klp library.