Move yast2 lan tests to YaST job group
Move yast2 lan tests to YaST job group trying to match the closer test coverage
to the existing one and therefore excluding this test module from
the default schedule of main.pm mechanism for latest product to
have more control on its maintenance.

Currently when the module is schedule in other arch than x86_64
we are just opening it, so with this change we can apply the
detailed check for all architectures.