salt: Fix ignored exit code on test.ping call
The code was never correct because we got the `script_run` return value
evaluation messed up. See
https://openqa.opensuse.org/tests/1257363#step/salt/20 for an example
where the initial `test.ping` call failed but the the test ignored the
failure and recorded a "passed" result for the test module.

As we do not record a soft_failure here anymore we can just call the
loop without a surrounding condition.

Related progress issue: https://progress.opensuse.org/issues/66691