Click authenticate in PolicyKit authentication dialog
Up to GNOME 3.36, the PolKit auth dialog started with 'authenticate'
having the focus. for users with a password that is ok, as no harm
can happen by wrongly 'just' hitting enter.

From GNOME 3.36 on, the default is to 'cancel' the auth request, so
that the user can't just 'press away an auth dialog without thought',
since the following actions could have a negative impact on the system.

So instead of pressing 'ret' after typing the password (blank password),
we explicitly click using the mouse.

https://progress.opensuse.org/issues/66083