Public Cloud: Able to reboot and wait for ssh with different username
In some cases, like migration scenarios, after triggering the
migration command, we need to reboot and the system will boot
and expect a different username for login than the default one.
This way, we can specify the username to the reboot command
so it will trigger the shutdown cmd in the PC instance and
wait_for_ssh with the needed username.