Check unused modules for changed files only
We always run this check for all modules in the repo, which takes around
5 minutes. However, we can just check files which are changed in the
  repo.
For that we have 3 cases:
1. Module is added or renamed
2. Last usage is removed from main.pm
3. Last usage is removed from any schedule file

With this change, we get modules from each of the group and verify if
they are used anywhere else and fail CI check in case it's not.

See [poo#64688](https://progress.opensuse.org/issues/64688).