LTP: Reuse UEFI_baremetal_basic to install LTP
LTP_BAREMETAL=1 is needed to not wait on boot in install_ltp.pm (could
be also detected by IPXE, because both install_ltp_baremetal_mlx and
this aarch64 baremetal test which define INSTALL_LTP also define IPXE,
but it's safer not depend on it).

Also rename yaml to reflect current purpose (requires to rename both
YAML_SCHEDULE` in the testsuite and the name in job group setup).