Create a passwordless ssh key in HA cluster
Before bsc#1169581 crmsh used the default ssh root key (or created one)
for the HA connection. Crmsh now uses a dedicated key for this.

As some tests need to do some passwordless ssh connections we now need to
generate a pair. This was already done before in HA YaST test.