Fix race condition for yast_keyboard
When the `yast keyboard set` runs it executes loadkeys for all the tty.
when the type_string run the last command has not finished apparently.
And this explains why when the module runs successful the error about the
ttyS0 does not appear. increasing the type_string seems to work
but it doesnt seem the right solution. But also replacing the type_string
with the assert_script_run is problematic as the language in the device
has changed and the commands inside the assert_script_run are localed
as german characters.