Handle node 2 fencing in qnetd split brain scenario
With upcoming fixes to bsc#1170037, the split brain test with
qdevice/qnetd will require to handle the fencing of node 2 when
the split brain scenario is introduced by the test code. This
introduces changes in the job schedules and test modules so the
node 2 test is able to handle the fence, and also to verify the
cluster status after fencing. Additionally, STONITH is being
temporarily disabled to allow the test to verify the expected
configuration of qdevice/qnetd during the split brain scenario,
and then enabled to allow the cluster to handle the split brain,
along with a cleanup of the iptables rules introduced to cause
the split brain scenario.

Also part of this commit:
- Ensure promotable resource runs in node 1 so qnetd selects this
node during the split brain failure.
- Make qnetd server wait until testing finishes in cluster nodes