Move mouse on livecd to prevent screenlock appearing
The commit changes the behavior of the test module to move mouse,
cursor while installation is performed with Live CD. That is done
to prevent screenlock to be appeared.

Change is done in scope of work to stabilize the installation
test suite with Live CD installation on aarch64.

Related ticket: poo#66355