Add totest tag on registry url
In the case of SLE base container images, there is need to test all of
the images on the SUT. In order to achieve this, suse_container_urls
needs to be modified to accept the different image versions.

Also, currently SLE base images are being released using a two step
process, so it is needed to test and pull the images from the registry
that contain "totest" in their path.