wicked: Skip create_hdd_wicked job if HDD already exists
We would like to to avoid recreation of HDD if it already exists.
This will be used in an environment where we will run openqa as CI
tests.
We are not able to run isotovideo directly, as we need the
functionality of schedule and multimachine jobs.

With this hack, we are able to just `isos post` our CI jobs and do not
care if the HDD is already created or not.