Add workaround for bsc#1172738 on s390x
The commit adds workaround to select Language on License Agreement
step of Installer on s390x.

The issue is that "Next" button is triggered, even though it is not
in focus while selecting language on License Agreement screen with
"Enter" on s390x.

The workaround uses mouse click on the language item in dropdown
instead of pressing 'Enter'.