Virtualized Infiniband Testing with SR-IOV
In order to run virtualized tests for Infiniband, we need to create
those VMs. Instead of creating custom VM images, we just download the
right image from openQA, modify it and create as many images as we want
from it.
For retrieval of the image, we use openqa2vm and the baremetal support
service.

Variables you need to set:
BAREMETAL_SUPPORT_SERVER: set to full url including the port
GET_IMAGE_FROM_TEST: set to the name of the test that creates the
                     desired image. Defaults to create_hdd_textmode.
VM_MEM: set to the amount of memory you want to assign to the VMs.
        Defaults to 2048.
VM_CPU: number of CPU cores assigned to VMs. Default to 2.

It is assumed the baremetal host that runs this test already has a
network bridge configured, which is the default for our machines with
Infiniband links.
Also, this test only supports Mellanox ConnectX-5 for now.

Signed-off-by: Michael Moese <mmoese@suse.com>