Use persistent block devices paths in HA tests
Currently HA tests rely on block devices provided by an iSCSI server
to perform some of the required filesystem tests (sbd, cluster_md,
drbd, etc). These block devices are taken as LUNs from a list provided
by the iSCSI server, and then passed to realpath(1), which returns
the resolved paths (/dev/sd?). Problem with these resolved paths, is
that they are not necessarily persistent between different boots of
the system, which can lead to HA resources failing to start if the
resolved path changes after a fence; it can also lead to failures in
the HA bootstrap process if the resolved paths of a given LUN are
different between nodes. This commit introduces changes so the HA test
use the persistent paths/LUNs from /dev/disk/by-path/... supplied by
the support server.