Add Support for yast2-cluster in SLES+HA 12 & 15-SP2
Current version of tests/ha/yast_cluster_init needs to address some
changes made to yast2-cluster, namely there is the possibility of a
notice at startup which prompts to install dependencies, and the
"Cluster - Service" screen feature different shortcuts/hotkeys to
enable startup of the cluster.

Additionally, the "Cluster - Corosync Qdevice" screen is not present
in SLES versions older than 15-GM, so code was added so the module
also works in older OS versions.